Start RectEnvelope implementation
Similar to ordinary envelope but implemented using SphereLine and hence
works even for switch rotated almost vertically (our strategy of building
polyhedron for ordinary envelope produces invalid figure in this case due
to surfaces intersection). It ends with extruded cubes so it should be
easy to make socket for entire (plate + envelope) part and to make body
modular.