Refactor composition of a final part using Monoid
Final figure is composed from parts which has holes. The difficulty of
combining such parts is that due to interlapping we need to compose figures
and only then remove holes. To simplify this task we introduce
HollowFigure data structure which holds positive and negative parts of
a figure. HollowFigures can be composed as a Monoid and then rendered as
a difference to Scad.